fix consum diamond

+++ Server/GDouDiZhuServer/src/DDZRoomManager.cpp
@@ -10,7 +10,7 @@ IGameRoom* DDZRoomManager::createRoom(uint8_t nGameType)
 	return nullptr;
 }
 
-uint8_t DDZRoomManager::getDiamondNeed(uint8_t nGameType, uint8_t nLevel, ePayRoomCardType payType, uint16_t nSeatCnt )
+uint16_t DDZRoomManager::getDiamondNeed(uint8_t nGameType, uint8_t nLevel, ePayRoomCardType payType, uint16_t nSeatCnt )
 {
 //#ifdef _DEBUG
 //	return 0;
@@ -30,11 +30,11 @@ uint8_t DDZRoomManager::getDiamondNeed(uint8_t nGameType, uint8_t nLevel, ePayRo
 	// is aa true ;
 	if (ePayType_AA == payType)
 	{
-		uint8_t vAA[] = { 1 , 2 , 3 };
+		uint16_t vAA[] = { 1 , 2 , 3 };
 		return vAA[nLevel] * 10;
 	}
 
 	// 6,1 . 12.2 , 18. 3
-	uint8_t vFangZhu[] = { 3 , 6 , 9 };
+	uint16_t vFangZhu[] = { 3 , 6 , 9 };
 	return vFangZhu[nLevel] * 10;
 }
